redesign interface and utterance working fine with music (bug with start running)

+++ StartRunning/Person.m
@@ -16,7 +16,7 @@
 	return self;
 }
 
--(id)initWithGender:(NSString *)gender andHeight:(NSInteger)height andWeight:(NSInteger)weight{
+-(id)initWithGender:(NSString *)gender andHeight:(NSInteger)height andWeight:(double)weight{
 	self = [super init];
 	if (self) {
 		_gender=gender;
